factorize in a component the mealformitem
